// projects/isy-angular-widgets/src/lib/input-char/model/datentyp.ts
/**
 * Data types according to DIN 91379
 */
export enum Datentyp {
  /** Data type A. */
  DATENTYP_A = 'DATENTYP_A',

  /** Data type B. */
  DATENTYP_B = 'DATENTYP_B',

  /** Data type C. */
  DATENTYP_C = 'DATENTYP_C',

  /** Data type D. */
  DATENTYP_D = 'DATENTYP_D',

  /** Data type E. */
  DATENTYP_E = 'DATENTYP_E'
}